style: Separate clock dir
