+++ src/App.js
@@ -8,9 +8,9 @@ import LocationProvider from 'Contexts/Location'
 
 // Components
 import Whiteboard from 'Components/Whiteboard/Whiteboard'
-// import Clock from 'Components/Clock/Clock'
+import Clock from 'Components/Clock/Clock'
 import Weather from 'Components/Weather/Weather'
-import Map from 'Components/Map/Map'
+// import Map from 'Components/Map/Map'
 
 import { SectionWrapper } from './Styles'
 
@@ -23,7 +23,7 @@ const App = () => {
 				layout={[
 					{ i: 'whiteboard', x: 0, y: 0, w: 8, h: 8, static: true },
 					{ i: 'images', x: 0, y: 8, w: 8, h: 4 },
-					{ i: 'map', x: 8, y: 0, w: 4, h: 6, static: true },
+					{ i: 'clock', x: 8, y: 0, w: 4, h: 6, static: true },
 					{ i: 'weather', x: 8, y: 6, w: 4, h: 6 },
 				]}
 				className='layout'
@@ -36,8 +36,8 @@ const App = () => {
 					<Whiteboard />
 				</SectionWrapper>
 				<SectionWrapper key='images'>Images</SectionWrapper>
-				<SectionWrapper key='map'>
-					<Map />
+				<SectionWrapper key='clock'>
+					<Clock />
 				</SectionWrapper>
 				<SectionWrapper key='weather'>
 					<Weather />
