+++ src/Components/Clock/Analog/Styles.jsx
